Fixed bugs

+++ src/main/java/controllers/ViewCreationsController.java
@@ -14,7 +14,6 @@ import javafx.scene.control.Tab;
 import javafx.scene.control.TabPane;
 import javafx.scene.control.TreeItem;
 import javafx.scene.control.TreeView;
-import javafx.scene.layout.VBox;
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
 import javafx.scene.text.Text;
@@ -119,6 +118,8 @@ public class ViewCreationsController extends Controller{
 
 	@FXML
 	private void play() {
+		
+		stopAudioPlayer();
 
 		Tab tab = _tabPane.getSelectionModel().getSelectedItem();
 
@@ -136,7 +137,6 @@ public class ViewCreationsController extends Controller{
 
 			}
 		} else {
-			stopAudioPlayer();
 
 			// all audio files are leaves
 			if (_listOfAudio.getSelectionModel().getSelectedItem().isLeaf()) {
@@ -157,6 +157,8 @@ public class ViewCreationsController extends Controller{
 
 	@FXML
 	private void delete() {
+		stopAudioPlayer();
+		
 		Tab tab = _tabPane.getSelectionModel().getSelectedItem();
 
 		if (tab == _creationTab) { 
@@ -187,7 +189,7 @@ public class ViewCreationsController extends Controller{
 					
 					File audioFolder = new File(System.getProperty("user.dir") + File.separator + "bin" + File.separator
 							+ "audio" + File.separator + term);
-					if (audioFolder.listFiles().length == 0) {
+					if (audioFolder.exists() && audioFolder.listFiles().length == 0) {
 						audioFolder.delete();
 					}
 					_mainApp.displayViewCreationsScene();
@@ -217,6 +219,8 @@ public class ViewCreationsController extends Controller{
 	@FXML
 	private void mainMenu() {
 
+		stopAudioPlayer();
+		
 		_mainApp.displayMainMenuScene();
 
 	}
